# kbmk_backend-main/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'KBMK UNMUL API Documentation'
  description: 'Dokumentasi API untuk team Frontend'
  contact:
    email: admin@admin.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: /api
paths:
  /register:
    post:
      tags:
        - Authentication
      summary: 'Mendaftarkan user baru'
      description: 'Endpoint untuk membuat akun user baru'
      operationId: 5d4fa5d5607ed675ba93907e1fc94534
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - username
                - email
                - password
                - password_confirmation
                - roles
              properties:
                username:
                  type: string
                  maxLength: 50
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: johndoe@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                password_confirmation:
                  type: string
                  example: password123
                roles:
                  type: string
                  enum:
                    - admin
                    - superadmin
                  example: admin
              type: object
      responses:
        201:
          description: 'Registrasi berhasil'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Register success!'
                  data:
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        example: 1|abcdef123456...
                      token_type:
                        type: string
                        example: Bearer
                    type: object
                type: object
        422:
          description: 'Error validasi'
  /login:
    post:
      tags:
        - Authentication
      summary: 'Login user'
      description: 'Mendapatkan token autentikasi'
      operationId: 67d6d0fdd54fba132e1b6f55ca254183
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: johndoe@example.com
                password:
                  type: string
                  example: password123
              type: object
      responses:
        200:
          description: 'Login berhasil'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Login success!'
                  data:
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        example: 1|abcdef123456...
                      token_type:
                        type: string
                        example: Bearer
                    type: object
                type: object
        401:
          description: 'Email atau password salah'
  /logout:
    post:
      tags:
        - Authentication
      summary: 'Logout user'
      description: 'Menghapus token autentikasi yang sedang aktif'
      operationId: 28d48c13fca984400ac8de99f54ee13a
      responses:
        200:
          description: 'Logout berhasil'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Logout success!'
                type: object
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  /me:
    get:
      tags:
        - Authentication
      summary: 'Mendapatkan data user yang sedang login'
      description: 'Mengembalikan data user berdasarkan token yang diberikan'
      operationId: bc2032a379aa97707ccc61511897c120
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'User data fetched successfully'
                  data:
                    $ref: '#/components/schemas/User'
                type: object
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    User:
      required:
        - username
        - email
        - password
        - roles
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          maxLength: 50
          example: johndoe
        email:
          type: string
          format: email
          example: johndoe@example.com
        roles:
          type: string
          enum:
            - admin
            - superadmin
          example: admin
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer
...